First/quite working version after refactor with thread-based scroller. Only a mutex problem left.

+++ online/hypereeg/hypereeg-acq-daemon/iirfilter.h
@@ -4,14 +4,14 @@
 // Direct Form II Transposed Single-Sample IIR Filter Class
 class IIRFilter {
 public:
- IIRFilter(const std::vector<float>& bCoeffs,const std::vector<float>& aCoeffs)
-           : b(bCoeffs),a(aCoeffs),order(aCoeffs.size()-1),w(order,0.0f) {}
- float filterSample(float x) {
-  float wn=x;
+ IIRFilter(const std::vector<double>& bCoeffs,const std::vector<double>& aCoeffs)
+                            : b(bCoeffs),a(aCoeffs),order(aCoeffs.size()-1),w(order,0.0f) {}
+ double filterSample(double x) {
+  double wn=x;
   for (size_t i=0;i<order;++i) wn-=a[i+1]*w[i];
   wn/=a[0];
 
-  float yn=b[0]*wn;
+  double yn=b[0]*wn;
   for (size_t i=0;i<order;++i) yn+=b[i+1]*w[i];
 
   for (size_t i=order-1;i>0;--i) w[i]=w[i-1];
@@ -21,9 +21,9 @@ public:
  }
 
 private:
- std::vector<float> b,a;
+ std::vector<double> b,a;
  size_t order;
- std::vector<float> w;
+ std::vector<double> w;
 };
 
 // Usage Example inside your AcqThread
@@ -32,8 +32,8 @@ private:
 // for (size_t ch=0;ch<totalChannels;++ch) channelFilters.emplace_back(b,a);
 
 // Then for each sample:
-// float rawValue=tcpS.amp[ampIdx].data[chIdx];
-// float filteredValue=channelFilters[chGlobalIdx].filterSample(rawValue);
+// double rawValue=tcpS.amp[ampIdx].data[chIdx];
+// double filteredValue=channelFilters[chGlobalIdx].filterSample(rawValue);
 // tcpS.amp[ampIdx].data[chIdx] = filteredValue;
 
 // Proceed to push tcpS into your circular buffer as before.
